requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pandas==2.1.4
numpy==1.24.3
python-multipart==0.0.6
openpyxl==3.1.2 
PyJWT